Connection tree code moved into connections.
Moving toward building a full connection map.
TODO Provide methods for connections to provide their own reductions.
